fix(todo-intent): split panel module and remove try-pass parse
